Rename to zaptest/tlogger
